Allow any order on input.
Before this change, we required our users to input files in the right
alphabetical sorting, in order to be able to process the files on the
fly. But this complicates the interfaces a lot for a memory optimization
that is irrelevant for most cases.

After this change we'll simply collect the files to a map.