add pgp signing interface.
This allows library callers to give us a gpg signing function, which
will be used to sign the contents. The callers can create the signature
however way they want, for example by using golang's openpgp package
or by forking out to gpg. This mostly depends on managing the secret
material. If they want to use the user's keys I guess forking to gpg
is not too bad.

The test is not great: rpmsample was modified to add a signature, which
basically returns the same bytes all the time. This does not actually
validate if rpm based distros accept the key.