tar2rpm allow specifying a build time stamp via the cli
Add a build_time flag which allows specifying the build timestamp for
people who like that.
To use the current time, you can use:
` --build_time $(date +%s)`

I strongly recommend not doing that, because it makes the build change
every time and hurt your caching and reproducibility, but people seem to
need this.

Also change the default timestamp to be empty to avoid the 2042 time.

Fixes #43