Fix tests by conditionally signing.
Since #53 our tests are broken, because centos refuses to work with an unverifyable signed header. So we add a flag to sign only in the relevant test, and there use the nosignature flag to avoid centos trying to validate the signature.

This is another reminder that we must fix out CI testing soon.