fix: handle packager, vendor and others
Packager, vendor, URL, and Epoch are being added to the headers even if
they have no value and are not required.

This PR moves them to pointers, and add them if they are non-null
values.

Could also not add them if they have their default value, which would
not break anyone using the lib, but might be less clear, especially if
someone wants to set epoch to 0, for example...