tar2rpm: add filename '-' as an alias for stdin and stdout
When no filenames are given the informative message
"tar2pm: reading tar content from stdin, writing rpm data to stdout."
is printed to stderr. This can be suppressed by explicitly
specifying '-' for TARFILE and RPMFILE (if only TAR- or RPMFILE is
given the message reflects that).

Also updated usage() to reflect that, and rename FILE to
TARFILE/RPMFILE.